moved all css files to styles folder
