+++ src/components/login-page.js
@@ -3,8 +3,8 @@ import { connect } from 'react-redux';
 import { Link, Redirect } from 'react-router-dom';
 
 import LoginForm from './login-form';
-import './login-page.css';
 import tplogo from './images/logo/TP-logo.jpg';
+import './styles/login-page.css';
 
 class LoginPage extends React.Component {
     // If we are logged in (which happens automatically when registration
