+++ src/components/registration-form.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Field, reduxForm, focus } from 'redux-form';
-import { Link, Redirect } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 
 import { registerUser } from '../actions/users';
 import { login } from '../actions/auth';
