+++ src/components/styles/landing-page.css
