+++ src/components/styles/registration-form.css
