+++ src/components/styles/registration-page.css
